Move Seam extensions into seam.js
