+++ src/lib/blueprint.test.ts
@@ -6,9 +6,12 @@ import {
   getSemanticMethod,
   getWorkspaceScope,
   type Method,
-  type OpenapiAuthMethod,
 } from 'lib/blueprint.js'
-import type { OpenapiOperation, OpenapiSchema } from 'lib/openapi/types.js'
+import type {
+  OpenapiAuthMethod,
+  OpenapiOperation,
+  OpenapiSchema,
+} from 'lib/openapi/types.js'
 
 test('createProperties: assigns appropriate default values', (t) => {
   const minimalProperties = {
