+++ src/lib/blueprint.ts
@@ -12,7 +12,6 @@ import type {
 import { findCommonOpenapiSchemaProperties } from './openapi/find-common-openapi-schema-properties.js'
 import { flattenOpenapiSchema } from './openapi/flatten-openapi-schema.js'
 import {
-  type AuthMethodSchema,
   EventResourceSchema,
   OpenapiOperationSchema,
   PropertySchema,
@@ -20,11 +19,17 @@ import {
 } from './openapi/schemas.js'
 import type {
   Openapi,
+  OpenapiAuthMethod,
   OpenapiOperation,
   OpenapiPathItem,
   OpenapiPaths,
   OpenapiSchema,
 } from './openapi/types.js'
+import {
+  mapOpenapiToSeamAuthMethod,
+  type SeamAuthMethod,
+  type SeamWorkspaceScope,
+} from './seam.js'
 
 export interface Blueprint {
   title: string
@@ -93,15 +98,6 @@ export interface Endpoint {
   authMethods: SeamAuthMethod[]
   workspaceScope: SeamWorkspaceScope
 }
-
-export type SeamAuthMethod =
-  | 'api_key'
-  | 'personal_access_token'
-  | 'console_session_token'
-  | 'client_session_token'
-  | 'publishable_key'
-
-export type SeamWorkspaceScope = 'none' | 'optional' | 'required'
 
 interface BaseParameter {
   name: string
@@ -571,6 +567,7 @@ const createEndpointFromOperation = async (
   const endpointAuthMethods = operationAuthMethods
     .map(mapOpenapiToSeamAuthMethod)
     .filter((authMethod): authMethod is SeamAuthMethod => authMethod != null)
+
   const workspaceScope = getWorkspaceScope(operationAuthMethods)
 
   const endpoint: Omit<Endpoint, 'codeSamples'> = {
@@ -606,8 +603,6 @@ const createEndpointFromOperation = async (
   }
 }
 
-export type OpenapiAuthMethod = z.infer<typeof AuthMethodSchema>
-
 export const getWorkspaceScope = (
   authMethods: OpenapiAuthMethod[],
 ): SeamWorkspaceScope => {
@@ -639,24 +634,6 @@ export const getWorkspaceScope = (
   if (hasOnlyScopedAuth) return 'required'
 
   return 'none'
-}
-
-type KnownOpenapiAuthMethod = Exclude<OpenapiAuthMethod, 'unknown'>
-
-const mapOpenapiToSeamAuthMethod = (
-  method: string,
-): SeamAuthMethod | undefined => {
-  const AUTH_METHOD_MAPPING: Record<KnownOpenapiAuthMethod, SeamAuthMethod> = {
-    api_key: 'api_key',
-    pat_with_workspace: 'personal_access_token',
-    pat_without_workspace: 'personal_access_token',
-    console_session_token_with_workspace: 'console_session_token',
-    console_session_token_without_workspace: 'console_session_token',
-    client_session: 'client_session_token',
-    publishable_key: 'publishable_key',
-  } as const
-
-  return AUTH_METHOD_MAPPING[method as KnownOpenapiAuthMethod]
 }
 
 export const createRequest = (
